+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * **Breaking Change**
   * Generic Template now expects an object instead of an array of elements
   * **New Feature**
+   * Support for Attachment Upload API
    * Added support for `pre_checkout` event
 
 ## 4.0.0